Move new crafting-system to crafting package
